% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{adj.matrix}
\alias{adj.matrix}
\alias{adj.matrix.dbn}
\alias{adj.matrix.dbn.fit}
\title{Get the adj matrix from a network}
\usage{
adj.matrix(x, ...)

\method{adj.matrix}{dbn}(dbn, ...)

\method{adj.matrix}{dbn.fit}(dbn, ...)
}
\arguments{
\item{x}{a network object (\code{dbn}, \code{dbn.fit}).}

\item{...}{further arguments passed to the method.}
}
\value{
for \code{dbn} / \code{dbn.fit} a data.frame

for \code{dbn} / \code{dbn.fit} a data.frame

for \code{dbn} / \code{dbn.fit} a data.frame
}
\description{
S3 generic returning a \code{data.frame} with the adjecency matrix of the model,
For a \code{dbn} object it returns a data.frame with 0/1 identifing the edge,
whereas for a \code{dbn.fit} object it returns

S3 generic returning a \code{data.frame} with the adjecency matrix of the model,
For a \code{dbn} object it returns a data.frame with 0/1 identifing the edge,
whereas for a \code{dbn.fit} object it returns

S3 generic returning a \code{data.frame} with the adjecency matrix of the model,
For a \code{dbn} object it returns a data.frame with 0/1 identifing the edge,
whereas for a \code{dbn.fit} object it returns
}
\examples{
#' \dontrun{
dbn <- random.structure.dbn(c("A","B","C"), .5, .5, 1)
adj.matrix(dbn)
}
\dontrun{
fit <- fit_random_dbn(random.structure.dbn(c("A","B","C"), .5, .5, 1), "continuous")
adj.matrix(fit)
}
}
