% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{build_g0_df}
\alias{build_g0_df}
\title{Given a data.frame and a markov_order builds a data.frame for the initial conditions}
\usage{
build_g0_df(data, markov_order = 1, separator = "_", extend_g0 = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{markov_order}{= 1 markov order of the data frame}

\item{separator}{the separator to use when defining the new variable names (A{separator}i)}

\item{extend_g0}{= TRUE if TRUE include times 0 to (markov_order - 1) as variables; if FALSE only time 0}
}
\value{
data.frame with each time from 0 to (markov_order - 1) as a variable
}
\description{
Given a data.frame and a markov_order builds a data.frame for the initial conditions
}
\details{
for each variable excluding Sample_id and Time creates
variables A_0, ..., A_(k - 1) where k is the markov_order. If extend_g0 is
FALSE only the A_0 variables are created.
}
\examples{
shifted_df <- build_g0_df(data, 4)
}
