% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{build_shifted_df}
\alias{build_shifted_df}
\title{Given a data.frame and a markov_order builds a transition data.frame}
\usage{
build_shifted_df(data, markov_order = 1, separator = "-")
}
\arguments{
\item{data}{a data.frame}

\item{markov_order}{= 1 markov order of the data frame}

\item{separator}{= "-" separator to use when creating variable X_t-k}
}
\value{
data.frame shifted df
}
\description{
Given a data.frame and a markov_order builds a transition data.frame
}
\details{
for each variable excluding Sample_id and Time creates
variables X_t, ..., X_t-k where k is the markov_order
}
\examples{
shifted_df <- build_shifted_df(data, 4)
}
