% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{dbn.forecasting}
\alias{dbn.forecasting}
\title{Forecast next values from observations}
\usage{
dbn.forecasting(dbn, observations, timepoints)
}
\arguments{
\item{observations}{a list (or data.frame) containing the observations for each variable.
data.frames are coerced to a list internally; the forecast trajectory is grown
in-place via append(), which is a list operation.}

\item{timepoints}{number of timepoints to be forecasted}

\item{fitted_DBN}{an object of class 'dbn'}
}
\value{
a dataframe containing the forecasting
}
\description{
Forecast next values from observations
}
\examples{
dbn.forecasting(fitted_DBN, observations, timepoints)
}
