% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{dbn.prep.data}
\alias{dbn.prep.data}
\title{Preprocessing function for Structure Learning algorithm.}
\usage{
dbn.prep.data(
  data,
  markov_order = 1,
  blacklist = NULL,
  whitelist = NULL,
  allow_intraslice_edges = TRUE,
  allow_t_0_edges = TRUE,
  extend_g0 = TRUE
)
}
\arguments{
\item{data}{object of type data.frame to be given as an input to structure learning algorithm}

\item{markov_order}{markov order to build the dataset}

\item{blacklist}{matrix with 'from' and 'to' columns of forbidden edges in the Dynamic Bayesian Network (default is NULL)}

\item{whitelist}{matrix with 'from' and 'to' columns of fixed edges in the Dynamic Bayesian Network (default is NULL)}

\item{allow_intraslice_edges}{= TRUE bool wether to allow edges of the form X_t -> Y_t}

\item{allow_t_0_edges}{= TRUE bool wether to allow edges of the form X_0 -> Y_0}

\item{extend_g0}{= TRUE bool weather to include in G_0 nodes with time < markov_order. If false G_0 will only have _0 nodes}
}
\value{
character vector made of ordered data to fit G_0, data to fit G_transition, blacklist to fit G_0, blacklist to fit G_transition, whitelist to fit G_0, whitelist to fit G_transition
}
\description{
Preprocessing function for Structure Learning algorithm.
}
\examples{
DBN <- random.structure.dbn(c("A", "B", "C"), 0.5, 0.5, markov_order = 1)
DBN_fitted <- generate_dbn_nodes_distributions(DBN, c("A", "B", "C"), TRUE, 2)
data <- dbn_sampling(DBN_fitted, 2000, 5)
dbn.prep.data(data)
}
