% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{dbn.sampling.R}
\alias{dbn.sampling.R}
\title{Generate a sampling dataset}
\usage{
dbn.sampling.R(fitted_dbn, n_samples, max_time)
}
\arguments{
\item{n_samples}{number of samples}

\item{max_time}{time series length}

\item{fitted_DBN}{an object of class 'dbn'}
}
\value{
the generated dataframe
}
\description{
Generate a sampling dataset
}
\examples{
dbn.sampling(DBN_example, N_samples, Time)
}
