% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling_cpp.R
\name{dbn.sampling}
\alias{dbn.sampling}
\title{Generate a sampling dataset}
\usage{
dbn.sampling(fitted_dbn, n_samples, max_time)
}
\arguments{
\item{fitted_dbn}{an object of class 'dbn.fit'}

\item{n_samples}{number of samples}

\item{max_time}{time series length}
}
\value{
the generated dataframe
}
\description{
Generate a sampling dataset
}
\examples{
dbn.sampling(DBN_example, N_samples, Time)
}
