% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_conversion.R
\name{get.g0.net}
\alias{get.g0.net}
\title{Transformation of a DBN in a G_0 network}
\usage{
get.g0.net(dbn)
}
\arguments{
\item{DBN}{object of class 'dbn' or "dbn.fit"}
}
\value{
object of class 'bn'
}
\description{
Transformation of a DBN in a G_0 network
}
\examples{
G_0 <- from_DBN_to_G_0(DBN_example)
}
