% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_conversion.R
\name{get.transition.net}
\alias{get.transition.net}
\title{Transformation of a DBN in a G_transition network}
\usage{
get.transition.net(dbn)
}
\arguments{
\item{DBN}{object of class 'dbn' or "dbn.fit"}
}
\value{
object of class 'bn'
}
\description{
Transformation of a DBN in a G_transition network
}
\examples{
G_transition <- get.transition.net(DBN_example)
}
