% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_children_set}
\alias{get_children_set}
\title{Get the children of a node}
\usage{
get_children_set(dbn, variable)
}
\arguments{
\item{G}{A DBN/dbn.fit/bn.fit object}

\item{n}{name of node n}
}
\value{
childrens of the node n.
}
\description{
Get the children of a node
}
\examples{
get_children_set(G, "A_t")
}
