% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll.R
\name{get_node_edges}
\alias{get_node_edges}
\title{get_node_edges}
\usage{
get_node_edges(dbn_transition, node, time_slice)
}
\arguments{
\item{dbn_transition}{a dbn transition}

\item{node}{name of the node in the dbn transition}

\item{time_slice}{an integer representing a time slice}
}
\value{
all the edges of a specific node in a time slice
}
\description{
This function gets all the edges of a node
}
