% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parent_set}
\alias{get_parent_set}
\title{Get the parents of a node}
\usage{
get_parent_set(dbn, variable)
}
\arguments{
\item{G}{A DBN/dbn.fit/bn.fit object}

\item{n}{name of node n}
}
\value{
parents of the node n.
}
\description{
Get the parents of a node
}
\examples{
get_parent_set(G, "A_t")
}
