% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variable_name}
\alias{get_variable_name}
\title{Get the name of a variable}
\usage{
get_variable_name(n)
}
\arguments{
\item{n}{variable name like (A_t-1)}
}
\value{
characther the name of the variable (A)
}
\description{
Get the name of a variable
}
\details{
equivalent to \code{split_variable_name(n)$name}
}
\examples{
name <- split_variable_name("A_t-1")
name == "A"
}
