% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variable_time}
\alias{get_variable_time}
\title{Get the time of a variable}
\usage{
get_variable_time(n)
}
\arguments{
\item{n}{variable name like (A_0)}
}
\value{
characther the time of the variable (t_0)
}
\description{
Get the time of a variable
}
\details{
equivalent to \code{split_variable_name(n)$time}
}
\examples{
time <- split_variable_time("A_0")
time == "t_0"
}
