% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variable_time_index}
\alias{get_variable_time_index}
\title{Get the time index of a node}
\usage{
get_variable_time_index(variable)
}
\arguments{
\item{n}{name of node}
}
\value{
time index of a node n
}
\description{
Get the time index of a node
}
\examples{
time_lag <- get_variable_time_index("A_t-1")
time_lag == 1
}
