% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{mmhc.dbn}
\alias{mmhc.dbn}
\title{Learn the equivalence class of DBN via Max-Min Hill Climbing algorithm}
\usage{
mmhc.dbn(
  data,
  markov_order = 1,
  score = "auto",
  test = "auto",
  max.sx = NULL,
  blacklist = NULL,
  whitelist = NULL,
  extend_g0 = TRUE,
  ...
)
}
\arguments{
\item{data}{a data.frame to be given as an input the to structure learning algorithm}

\item{score}{a character string, the label of the scoring function to be used in the algorithm. Default \code{"auto"} selects the score from the data type via \code{\link[=dataset_type]{dataset_type()}}: \code{bic} (discrete), \code{bic-g} (gaussian) or \code{bic-cg} (mixed)}

\item{test}{a character string, the label of the conditional independence test to be used in the algorithm. Default \code{"auto"} selects the test from the data type via \code{\link[=dataset_type]{dataset_type()}}: \code{mi} (discrete), \code{cor} (gaussian) or \code{mi-cg} (mixed)}

\item{max.sx}{an integer, the maximum number of parents for each node (default is no limit)}

\item{blacklist}{a matrix with columns 'from' and 'to' defining the set of arcs that are forbidden}

\item{whitelist}{a matrix with columns 'from' and 'to' defining the set of arcs that are fixed}

\item{extend_g0}{= TRUE bool whether to include in G_0 nodes with time < markov_order. If false G_0 will only have _0 nodes}

\item{...}{additional parameters of the scoring function and/or chosen test}

\item{markov_oder}{an integer refering to the markov order of the learning algorithm}
}
\value{
An object of class 'dbn', which is a DBN learned via  Max-Min Hill Climbing algorithm
}
\description{
Learn the equivalence class of DBN via Max-Min Hill Climbing algorithm
}
\examples{
DBN <- random.structure.dbn(c("A", "B", "C"), 0.5, 0.5, markov_order = 1)
DBN_fitted <- generate_dbn_nodes_distributions(DBN, c("A", "B", "C"), TRUE, 2)
data <- dbn_sampling(DBN_fitted, 2000, 5)
mmhc.dbn(data)

}
