% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{print.dbn.fit}
\alias{print.dbn.fit}
\title{Print a fitted Dynamic Bayesian Network}
\usage{
\method{print}{dbn.fit}(x, ...)
}
\arguments{
\item{x}{object of class 'dbn.fit'}

\item{...}{further arguments passed to or from other methods}
}
\value{
the input \code{x} object, invisibly
}
\description{
Print a fitted Dynamic Bayesian Network
}
\examples{
print(fitted_dbn)
fitted_dbn
}
