% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{print.dbn}
\alias{print.dbn}
\title{Print a Dynamic Bayesian Network}
\usage{
\method{print}{dbn}(DBN, ...)
}
\arguments{
\item{DBN}{object of class 'dbn'}

\item{...}{further arguments passed to or from other methods}
}
\value{
the input \code{DBN} object, invisibly
}
\description{
Print a Dynamic Bayesian Network
}
\examples{
DBN_example <- empty.dbn(static_nodes = c(), dynamic_nodes = c("A", "B", "C"), markov_order = 1)
print(DBN_example)
DBN_example
}
