% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_conversion.R
\name{redefine_class}
\alias{redefine_class}
\title{This takes a fit object and renames the class names from the dbn to bn,
both for each node and for the network itself.}
\usage{
redefine_class(net)
}
\arguments{
\item{net}{a fitted network (list of nodes)}
}
\value{
the network with bn.fit classes
}
\description{
Nodes are relabeled from 'dbn.fit.{cg,d,g}node' to 'bn.fit.{cg,d,g}node'.
The network class is set to c("bn.fit", "bn.fit.<net_type>") where net_type
is 'dnet' (all discrete nodes), 'gnet' (all Gaussian nodes) or 'cgnet'
(mixed / conditional Gaussian nodes).
}
