% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll.R
\name{rename_nodes_unroll}
\alias{rename_nodes_unroll}
\title{rename_nodes_unroll}
\usage{
rename_nodes_unroll(time_slice, nodes_names)
}
\arguments{
\item{time_slice}{an integer representing a time slice}

\item{nodes_names}{a list of nodes names}
}
\description{
Renames a list of nodes, making nodes names compliant to the dbn unrolled structure.
A node at time \code{t} (\code{X_t}) is mapped to \code{X_{time_slice}} and a lagged
node \code{X_t-k} (any Markov lag \eqn{k \ge 1}) is mapped to \code{X_{time_slice - k}}.
}
