% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_variable_name}
\alias{split_variable_name}
\title{Split the name of a variable in a list with name and time}
\usage{
split_variable_name(variable)
}
\arguments{
\item{n}{name of node}
}
\value{
list containing name and time
}
\description{
Split the name of a variable in a list with name and time
}
\examples{
var_split <- split_variable_name("A_t-1")
var_split == ("A", "t-1")
}
