% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{add.node.dbn}
\alias{add.node.dbn}
\title{Function for node addiction in Dynamic Bayesian Network}
\usage{
add.node.dbn(DBN, node, type = "Dynamic")
}
\arguments{
\item{DBN}{object of class 'dbn'}

\item{node}{object of class 'character': name of the node to add}

\item{type}{'Dynamic' (default) or 'Static' node}
}
\value{
object of class 'dbn'
}
\description{
Function for node addiction in Dynamic Bayesian Network
}
\examples{
DBN_example <- add.node.dbn(DBN=DBN_example, node="F", type='Dynamic')
DBN_example <- add.node.dbn(DBN=DBN_example, node="K", type='Static')
}
