% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_helpers.R
\name{blacklist_g0_gt}
\alias{blacklist_g0_gt}
\title{create blacklist form list of nodes in dbn}
\usage{
blacklist_g0_gt(
  g_0_nodes,
  g_t_nodes,
  markov_order = 1,
  static_nodes = c(),
  allow_intraslice_edges = TRUE,
  allow_t_0_edges = TRUE
)
}
\arguments{
\item{g_0_nodes}{all nodes in the g_0 network like: "A_0", ...}

\item{g_t_nodes}{all nodes ending in t like: "A_t", ...}

\item{markov_order}{markov_order to generate blacklist from}

\item{static_nodes}{character vector of time-independent node names (without suffix), e.g. c("S", "K")}

\item{allow_intraslice_edges}{Default TRUE weather or not to allow intraslice edges (A_t -> B_t)}

\item{allow_t_0_edges}{Default TRUE weather or not to allow edges at t = 0 (A_0 -> B_0)}
}
\value{
matrix of node pairs which edge is forbidden
}
\description{
create blacklist form list of nodes in dbn
}
\examples{
blacklist_g0_gt(c("A_0", "B_0"), c("A_t", "B_t"), markov_order = 3)
blacklist_g0_gt(c("A_0", "B_0"), c("A_t", "B_t"), static_nodes = c("S"))

}
