% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_parameter_learning.R
\name{dbn.fit}
\alias{dbn.fit}
\title{Function for parameter learning in Dynamic Bayesian Networks}
\usage{
dbn.fit(
  DBN,
  CPTs = list(),
  CPDs = list(),
  data = data.frame(),
  replace.unidentifiable = FALSE
)
}
\arguments{
\item{DBN}{object of class 'dbn'}

\item{CPTs}{list of multi-dimensional arrays (CPTs for each DBN node, discrete)}

\item{CPDs}{list of named numeric vectors (regression coefficients + std, gaussian)}

\item{data}{data.frame object}

\item{replace.unidentifiable}{If TRUE conditional probabilities for unobserved
parents combinations (unidentifiable parameters) are replaced by uniform
conditional probabilities, if FALSE (default) they are set as NA. Discrete only.}
}
\value{
object of class 'dbn.fit'
}
\description{
If a data.frame is given, the function inspects the class of each column exept
\code{Time} and \code{Sample_id}, if all columns are factor discrete parameter learning is used
if they are all numeric gaussian parameter learning. Mixed case is not supported at the time.
}
\details{
Dispatcher for parameter learning. Routes to the discrete or gaussian
subroutine based on what the user provides:
\itemize{
\item \code{CPTs} non-empty -> discrete network learned from CPTs
\item \code{CPDs} non-empty -> gaussian network learned from CPDs
\item only \code{data}      -> dataset_type(data) decides the subroutine
}
Exactly one of \code{CPTs}, \code{CPDs} or \code{data} must be non-empty.
}
\examples{
learned_dbn <- dbn.fit(DBN = DBN_example, data = sampling_set)
}
