% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbn_type}
\alias{dbn_type}
\title{Return type of dbn}
\usage{
dbn_type(dbn)
}
\arguments{
\item{dbn}{object of type dbn.fit}
}
\value{
string "discrete", "gaussian" or "mixed"
}
\description{
Return type of dbn
}
\details{
if the type is not recognized raise error
}
\examples{
type <- dbn_type(dbn)
}
