% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{empty.dbn}
\alias{empty.dbn}
\title{Function to create a Dynamic Bayesian Network}
\usage{
empty.dbn(static_nodes = c(), dynamic_nodes, markov_order)
}
\arguments{
\item{static_nodes}{character vector listing the time-independent nodes}

\item{dynamic_nodes}{character vector listing the time-dependent nodes}

\item{markov_order}{integer value (> 1): markovian order of the process}
}
\value{
object of class 'dbn'
}
\description{
Function to create a Dynamic Bayesian Network
}
\examples{
DBN_example <- empty.dbn(static_nodes = c("K"), dynamic_nodes = c("A", "S", "E", "O", "R", "T", "B"), markov_order = 1)
class(DBN_example) # 'dbn'
}
