% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_mo_dbn_fit}
\alias{get_max_mo_dbn_fit}
\title{Return the maximum markov order of a dbn.fit object}
\usage{
get_max_mo_dbn_fit(dbn)
}
\arguments{
\item{dbn}{an object of class dbn.fit}
}
\value{
markov order of the dbn
}
\description{
Return the maximum markov order of a dbn.fit object
}
\examples{
mo <- get_max_mo_dbn_fit(fit)
}
